FEATURE: moves Updater class into app folder
