+++ libs/spirit-islander/util/src/index.ts
@@ -1,5 +1,3 @@
-export * from './lib/updater';
-
 // Game data and types
 export * from './lib/game/adversaries';
 export * from './lib/game/boards';
@@ -24,3 +22,4 @@ export * from './lib/app/game-setup.interface';
 export * from './lib/app/get-valid-combos';
 export * from './lib/app/route.enum';
 export * from './lib/app/select-boards';
+export * from './lib/app/updater';
