+++ libs/spirit-islander/util/src/lib/app/updater.spec.ts
